==> src/1_pipeline_download_reddit_comments.ipynb <==
# Descargar comentarios desde reddit

En este *notebook* se descargan comentarios desde reddit argentina.

### Fuente

- [reddit-api](https://www.jcchouinard.com/reddit-api-without-api-credentials/)

import praw
import pandas as pd
import os

TEXT_SAVE_FILE = 'docs/reddit_data.csv'
MIN_WORDS_COMMENTS = 5
MAX_WORDS_COMMENTS = 40
SUBMISSION_LIMIT = 50000

Realizamos la conexión a Reddit con las credenciales de la API que tengamos.

reddit = praw.Reddit(client_id=os.environ['REDDIT_CLIENT_ID'],
                     client_secret=os.environ['REDDIT_CLIENT_SECRET'],
                     user_agent=os.environ['REDDIT_USER_AGENT'])



# Objetivo

En este paso, extraemos los comentarios de reddit Argentina.

Cada *post* en reddit esta conformado por comentarios, y a la vez cada comentario puede tener respuestas. Nos interesa los comentarios en diferentes niveles, tanto los que pertenecen al *post* como los que son respuestas a otros comentarios.

Los comentarios en reddit pueden ser *link* o pueden ser solo textos. Filtramos solamente los comentarios que tengan textos. Para esto obtenemos cada comentario en formato *markdown* y eliminamos aquellos que comiencen con **>**.

## Estructura de datos

De cada comentario que se guarde de reddit, se obtendrá con los siguientes datos:

- *score*: es un puntaje que los usuarios le dan al comentario.
- *id*: identificador de reddit.
- *flair*: es el tipo de comentario etiquetado por reddit, por ejemplo, política, economía, humor, etc.
- *comms_num*: número de respuestas que recibió el comentaio.
- *comment_parent_id*: identificador del *post* o comentario al cual responde el comentario actual.
- *is_replay*: si es *False* significa que es un comentario de un post, en caso que sea *True* significa que es una respuesta a otro comentario.

topics_dict = { "score":[], "id":[], "flair":[], \
                "comms_num": [],  "body":[], "comment_parent_id":[], "is_replay": []}

for submission in reddit.subreddit('argentina').new(limit=SUBMISSION_LIMIT):
    submission_comm = reddit.submission(id=submission.id)

    # replication of the comment section of the reddit post
    def replies_of(top_level_comment, link_flair_text, topics_dict):
        if len(top_level_comment.replies) == 0:
            return
        for num, comment in enumerate(top_level_comment.replies):
            stringCleanComment = depure(comment.body)
            if not stringCleanComment == '':
                #print('-' * 5, stringCleanComment)
                topics_dict["flair"].append(link_flair_text)
                topics_dict["score"].append(comment.score)
                topics_dict["id"].append(comment.id)
                topics_dict["is_replay"].append(True)
                topics_dict["comms_num"].append(len(comment.replies))
                topics_dict["comment_parent_id"].append(top_level_comment.id)
                topics_dict["body"].append(depure(comment.body))
                if len(comment.replies) == 0:
                    return
                replies_of(comment, link_flair_text, topics_dict)

    def depure(text):
        stringClean = ""
        for cline in text.splitlines( ):
            line = cline.strip()
            if not line.startswith('>') and len(line) < 1500 and not line.startswith('#####&#009') and not line.startswith('######&#009;')  and not line.startswith('[Owner]')  and not line.startswith('####&#009;') and not line == '' and not line.startswith('- - - - - -'): 
                if stringClean == '':
                    stringClean = line
                else:
                    stringClean = stringClean + '. ' + line
        countWords = len(stringClean.strip().split(" "))
        if countWords > MIN_WORDS_COMMENTS and countWords < MAX_WORDS_COMMENTS:
            return stringClean
        else:
            return ''

    for count, top_level_comment in enumerate(submission_comm.comments):
        count_comm = 0
        try :
            stringCleanComment = depure(top_level_comment.body)
            if not stringCleanComment == '':
                topics_dict["flair"].append(submission.link_flair_text)
                topics_dict["score"].append(top_level_comment.score)
                topics_dict["id"].append(top_level_comment.id)
                topics_dict["is_replay"].append(False)
                topics_dict["comms_num"].append(len(top_level_comment.replies))
                topics_dict["comment_parent_id"].append(submission.id)
                topics_dict["body"].append(stringCleanComment)
            replies_of(top_level_comment, submission.link_flair_text, topics_dict)
        except:
            continue



# Objetivo

Se guardan los datos obtenidos en el paso anterior en un archivo csv.

topics_data = pd.DataFrame(topics_dict)

topics_data.to_csv(TEXT_SAVE_FILE, index=False)


print(topics_data.head(10))
