==> src/3b_pipeline_embedding_word2vec.ipynb <==
# Embeddings Neuronales

En este notebook, se utiliza la librería [Word2vec](https://en.wikipedia.org/wiki/Word2vec) como modelo de generación de embeddings de palabras, luego se entrena un modelo k-means para obtener clusters, y se marca cada comentario con un determinado número de cluster.

### Importación de librerías requeridas

import os
import pickle
import warnings

import gensim.corpora as corpora
from gensim.models import Word2Vec

from collections import Counter
import numpy as np

import pandas as pd

from sklearn.cluster import KMeans
from sklearn.metrics import pairwise_distances, silhouette_samples, silhouette_score

from clustering_utils import vectorize, mbkmeans_clusters

warnings.filterwarnings("ignore", category=DeprecationWarning)

np.random.seed(42)

%matplotlib inline


### Definición de variables globales

TEXT_FILE_READ = 'docs/preprocessing_reddit_data.csv'
TEXT_SAVE_FILE = 'docs/reddit_data_word2vec.csv'
PICKLE_KMEANS = 'docs/models/word2vec_kmeans.model'
TEST_CLUSTER_PATH = 'docs/test/w2v_comments_per_cluster/'
SAVE_TEST_CSV=False

n_clusters = 120

### Lectura de los comentarios de Reddit

Los comentarios fueron previamente preprocesados (Ver en TODO).

df = pd.read_csv(TEXT_FILE_READ)
df['lemma_tokens'] = df['lemma_tokens'].apply(eval)

### Vocabulario

# Create Dictionary
id2word = corpora.Dictionary(df['lemma_tokens'])

# Filtering Extremes
id2word.filter_extremes(no_below=2, no_above=.99)

# Creating a corpus object
corpus = [id2word.doc2bow(d) for d in df['lemma_tokens']]


processed_corpus = df['lemma_tokens']


### Entrenamiento del modelo Word2Vec

model = Word2Vec(sentences=processed_corpus, vector_size=100, window=5, min_count=1, workers=1)
model.train(processed_corpus, total_examples=len(processed_corpus), epochs=100)
model.save("docs/models/word2vec.model")

word_vecs = []
vocabulary = list(model.wv.key_to_index)

for key in model.wv.key_to_index:
    word_vecs.append(model.wv[key])

### Generación de vectores desde documentos

vectorized_docs = vectorize(processed_corpus, model=model)
len(vectorized_docs), len(vectorized_docs[0])

### Generación de clusters

clustering, cluster_labels = mbkmeans_clusters(
    X=vectorized_docs,
    k=n_clusters,
    mb=500,
    print_silhouette_values=True,
)
df_clusters = pd.DataFrame({
    "text": df["body"].values,
    "tokens": [" ".join(text) for text in processed_corpus],
    "cluster": cluster_labels
})

with open(PICKLE_KMEANS, 'wb') as f:
    pickle.dump(clustering, f)

### *Top terms* por cluster (basado en los centroides de los clusters)

print("Most representative terms per cluster (based on centroids):")
for i in range(n_clusters):
    tokens_per_cluster = ""
    most_representative = model.wv.most_similar(positive=[clustering.cluster_centers_[i]], topn=5)
    for t in most_representative:
        tokens_per_cluster += f"{t[0]} "
    print(f"Cluster {i}: {tokens_per_cluster}")



### *Top terms* por cluster (basado en las palabras más frecuentes)

for i in range(n_clusters):
    tokens_per_cluster = ""
    most_frequent = Counter(" ".join(df_clusters.query(f"cluster == {i}")["tokens"]).split()).most_common(5)
    for t in most_frequent:
        tokens_per_cluster += f"{t[0]}({str(t[1])}) "
    print(f"Cluster {i}: {tokens_per_cluster}")


### Recupere los documentos más representativos (basados en los centroides de los clústeres) para un cluster en particular

test_cluster = 17
most_representative_docs = np.argsort(
    np.linalg.norm(vectorized_docs - clustering.cluster_centers_[test_cluster], axis=1)
)
for d in most_representative_docs[:20]:
    print( df["body"].values[d])
    print("-------------")

#solo test
#print(len(vectorized_docs))
#print(vectorized_docs[0])

test_v = vectorize([['defender', 'peso', 'siente', 'corazón', 'compro', 'pesos', 'tasa', 'fijo', 'año']], model=model)
prediction = clustering.predict(test_v)
print(prediction)

ver = "['defender', 'peso', 'siente', 'corazón', 'compro', 'pesos', 'tasa', 'fijo', 'año']"
ver = "tapastir baño tirar balde aguo"
test_v = vectorize([ver], model=model)
prediction = clustering.predict(test_v)
print(prediction)


str2 = ver.split(" ")
print(str2)

reddit = pd.read_csv(TEXT_FILE_READ)

def get_cluster(row):
    test_v = vectorize([str(row).split(" ")], model=model)
    return clustering.predict(test_v)

reddit['cluster'] = reddit.apply(lambda row: get_cluster(row['body_preprocessing']) , axis = 1) 


# Show
df.head(10)

df.to_csv(TEXT_SAVE_FILE, index=False)

if SAVE_TEST_CSV:
    os.makedirs(TEST_CLUSTER_PATH,exist_ok=True)

    for i in range(n_clusters):
        df[(df["cluster"] == i)][['flair', 'body']].to_csv(TEST_CLUSTER_PATH + str(i) + '.csv')
